cursor sort of works
