+++ gtars/src/uniwig/counting.rs
@@ -9,7 +9,7 @@ use noodles::sam::alignment::Record;
 use std::collections::HashMap;
 use std::fs::{create_dir_all, File, OpenOptions};
 use std::io;
-use std::io::{stdout, BufRead, BufReader, BufWriter, Write};
+use std::io::{stdout, BufRead, BufReader, BufWriter, Cursor, Write};
 use tokio::runtime;
 
 /// This function is a more direct port of smoothFixedStartEndBW from uniwig written in CPP.
@@ -459,14 +459,15 @@ pub fn fixed_start_end_counts_bam_to_bw(
     chrom_size: i32,
     smoothsize: i32,
     stepsize: i32,
-    output_type: &str,
     chromosome_name: &String,
     bwfileheader: &str,
     out_sel: &str,
     std_out_sel: bool,
-    bedgraphargstruct: BedGraphToBigWigArgs,
-) {
+) -> Cursor<String> {
     //let vin_iter = starts_vector.iter();
+
+    //let mut vec_lines: Vec<String> = Vec::new();
+    let mut bedgraphlines = String::new();
 
     let mut v_coordinate_positions: Vec<i32> = Vec::new(); // these are the final coordinates after any adjustments
     let mut v_coord_counts: Vec<u32> = Vec::new(); // u8 stores 0:255 This may be insufficient. u16 max is 65535
@@ -497,45 +498,6 @@ pub fn fixed_start_end_counts_bam_to_bw(
 
     adjusted_start_site = adjusted_start_site - smoothsize;
 
-    // SET UP BW FILE WRITER HERE
-    //------------------------------------
-    let chrom_map: HashMap<String, u32> =
-        BufReader::new(File::open(bedgraphargstruct.chromsizes).unwrap())
-            .lines()
-            .filter(|l| match l {
-                Ok(s) => !s.is_empty(),
-                _ => true,
-            })
-            .map(|l| {
-                let words = l.expect("Split error");
-                let mut split = words.split_whitespace();
-                (
-                    split.next().expect("Missing chrom").to_owned(),
-                    split.next().expect("Missing size").parse::<u32>().unwrap(),
-                )
-            })
-            .collect();
-
-    let mut outb = BigWigWrite::create_file(bedgraphargstruct.bedgraph, chrom_map).unwrap();
-    outb.options.max_zooms = bedgraphargstruct.write_args.nzooms;
-    outb.options.compress = !bedgraphargstruct.write_args.uncompressed;
-    outb.options.input_sort_type = InputSortType::START;
-    outb.options.block_size = bedgraphargstruct.write_args.block_size;
-    outb.options.inmemory = bedgraphargstruct.write_args.inmemory;
-    let runtime = if bedgraphargstruct.write_args.nthreads == 1 {
-        outb.options.channel_size = 0;
-        runtime::Builder::new_current_thread().build().unwrap()
-    } else {
-        runtime::Builder::new_multi_thread()
-            .worker_threads(bedgraphargstruct.write_args.nthreads)
-            .build()
-            .unwrap()
-    };
-    let allow_out_of_order_chroms = !matches!(outb.options.input_sort_type, InputSortType::ALL);
-    //------------------------------------
-    //FINISHED SETTING UP BW WRITER
-    let stdin = std::io::stdin().lock();
-    let mut stdout = stdout().lock();
 
     current_end_site = adjusted_start_site;
     current_end_site = adjusted_start_site + 1 + smoothsize * 2;
@@ -582,88 +544,91 @@ pub fn fixed_start_end_counts_bam_to_bw(
             continue;
         }
 
-        while coordinate_position < adjusted_start_site {
-            while current_end_site == coordinate_position {
-                count = count - current_score;
+        //let single_line = format!("{}\t{}\t{}\t{}\n", chromosome_name, adjusted_start_site, current_end_site, count);
+        let single_line = format!("chr1\t{}\t{}\t2\n", adjusted_start_site, adjusted_start_site+2);
+        //let mut single_line = String::from("chr1\t1156063\t1156075\t2\n");
+        //vec_lines.push(single_line);
+        bedgraphlines.push_str(&*single_line);
+        println!("{}",bedgraphlines);
+        println!("iteration");
 
-                if count < 0 {
-                    count = 0;
-                }
 
-                if collected_end_sites.last() == None {
-                    current_end_site = 0;
-                } else {
-                    current_end_site = collected_end_sites.remove(0)
-                }
-            }
-
-            if coordinate_position % stepsize == 0 {
-                // Step size defaults to 1, so report every value
-                //v_coord_counts.push(count as u32);
-
-                // writeln!(
-                //     &mut buf,
-                //     "{}\t{}\t{}\t{}",
-                //     chromosome_name, adjusted_start_site, current_end_site, count
-                // )
-                //     .unwrap();
-                write!(
-                    stdout,
-                    "{}\t{}\t{}\t{}",
-                    chromosome_name, adjusted_start_site, current_end_site, count
-                )
-                .unwrap(); // THIS IS A FIXED STEP BEDGRAPH LINE
-
-                //v_coordinate_positions.push(coordinate_position);
-            }
-
-            coordinate_position = coordinate_position + 1;
-        }
-
-        prev_coordinate_value = adjusted_start_site;
+        // while coordinate_position < adjusted_start_site {
+        //     while current_end_site == coordinate_position {
+        //         count = count - current_score;
+        //
+        //         if count < 0 {
+        //             count = 0;
+        //         }
+        //
+        //         if collected_end_sites.last() == None {
+        //             current_end_site = 0;
+        //         } else {
+        //             current_end_site = collected_end_sites.remove(0)
+        //         }
+        //     }
+        //
+        //     if coordinate_position % stepsize == 0 {
+        //         // Step size defaults to 1, so report every value
+        //         //v_coord_counts.push(count as u32);
+        //
+        //         // writeln!(
+        //         //     &mut buf,
+        //         //     "{}\t{}\t{}\t{}",
+        //         //     chromosome_name, adjusted_start_site, current_end_site, count
+        //         // )
+        //         //     .unwrap();
+        //         // write!(
+        //         //     stdout,
+        //         //     "{}\t{}\t{}\t{}",
+        //         //     chromosome_name, adjusted_start_site, current_end_site, count
+        //         // )
+        //         // .unwrap(); // THIS IS A FIXED STEP BEDGRAPH LINE
+        //         let single_line = format!("{}\t{}\t{}\t{}",
+        //                                   chromosome_name, adjusted_start_site, current_end_site, count);
+        //         //v_coordinate_positions.push(coordinate_position);
+        //     }
+        //
+        //     coordinate_position = coordinate_position + 1;
+        // }
+        //
+        // prev_coordinate_value = adjusted_start_site;
     }
+    let mut cursor = Cursor::new(bedgraphlines);
+    cursor
 
-    count = count + 1; // We must add 1 extra value here so that our calculation during the tail as we close out the end sites does not go negative.
-                       // this is because the code above subtracts twice during the INITIAL end site closure. So we are missing one count and need to make it up else we go negative.
+    // count = count + 1; // We must add 1 extra value here so that our calculation during the tail as we close out the end sites does not go negative.
+    //                    // this is because the code above subtracts twice during the INITIAL end site closure. So we are missing one count and need to make it up else we go negative.
+    //
+    // while coordinate_position < chrom_size {
+    //     // Apply a bound to push the final coordinates otherwise it will become truncated.
+    //
+    //     while current_end_site == coordinate_position {
+    //         let current_score = adjusted_start_site;
+    //         count = count - current_score;
+    //         if count < 0 {
+    //             count = 0;
+    //         }
+    //
+    //         if collected_end_sites.last() == None {
+    //             current_end_site = 0;
+    //         } else {
+    //             current_end_site = collected_end_sites.remove(0)
+    //         }
+    //     }
+    //
+    //     if coordinate_position % stepsize == 0 {
+    //         // Step size defaults to 1, so report every value
+    //         //v_coord_counts.push(count as u32);
+    //         let single_line = format!("{}\t{}\t{}\t{}",
+    //                                   chromosome_name, adjusted_start_site, current_end_site, count);
+    //         //v_coordinate_positions.push(coordinate_position);
+    //     }
+    //
+    //     coordinate_position = coordinate_position + 1;
+    // }
 
-    while coordinate_position < chrom_size {
-        // Apply a bound to push the final coordinates otherwise it will become truncated.
 
-        while current_end_site == coordinate_position {
-            let current_score = adjusted_start_site;
-            count = count - current_score;
-            if count < 0 {
-                count = 0;
-            }
-
-            if collected_end_sites.last() == None {
-                current_end_site = 0;
-            } else {
-                current_end_site = collected_end_sites.remove(0)
-            }
-        }
-
-        if coordinate_position % stepsize == 0 {
-            // Step size defaults to 1, so report every value
-            //v_coord_counts.push(count as u32);
-
-            write!(
-                stdout,
-                "{}\t{}\t{}\t{}",
-                chromosome_name, adjusted_start_site, current_end_site, count
-            )
-            .unwrap(); // THIS IS A FIXED STEP BEDGRAPH LINE
-
-            //v_coordinate_positions.push(coordinate_position);
-        }
-
-        coordinate_position = coordinate_position + 1;
-    }
-    let vals = BedParserStreamingIterator::from_bedgraph_file(stdin, allow_out_of_order_chroms);
-
-    outb.write(vals, runtime).unwrap();
-    let _ = stdout.flush();
-    //println!("FInished with fixed_start_end_counts_bam");
     //(v_coord_counts, v_coordinate_positions)
 }
